ai-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.6.0
pycountry==24.6.1
google-genai>=0.8.0
google-adk>=0.1.2
